cdc_psql_kafka/requirements.txt
delta-spark==1.2.1
ipython-sql==0.4.0
psycopg2-binary==2.9.3